# perf: parse rego input once